Adapt `Ellipsoid` constructor to match type of keyword arguments
When no type is explicitely passed to the `Ellipsoid` constructor, it should take the promoted type only from the keyword arguments. Thus, the default parameter are defined as `Int`. If all parameters are `Int`, the `Ellipsoid` methods that perform the type promotion ensure that the `Ellipsoid` is always of type `Float`.

+++ src/ConstructiveSolidGeometry/VolumePrimitives/Ellipsoid.jl
@@ -65,14 +65,14 @@ end
 #Type promotion happens here
 function Ellipsoid(CO, r::TR, origin::PT, rotation::ROT) where {TR, PT, ROT}
     eltypes = _csg_get_promoted_eltype.((TR, PT, ROT))
-    T = promote_type(eltypes...)
+    T = float(promote_type(eltypes...))
     Ellipsoid{T,CO}(r, origin, rotation)
 end
 
 function Ellipsoid(::Type{CO}=ClosedPrimitive;
-    r = 1.0, 
-    origin = zero(CartesianPoint{Float64}), 
-    rotation = one(SMatrix{3, 3, Float64, 9})
+    r = 1, 
+    origin = zero(CartesianPoint{Int64}), 
+    rotation = one(SMatrix{3, 3, Int64, 9})
 ) where {CO}
     Ellipsoid(CO, r, origin, rotation)
 end
